% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGdata.R
\name{AGdata}
\alias{AGdata}
\title{Association study for American Gut project}
\usage{
AGdata(nOTU = 100, S1, S2, pat)
}
\arguments{
\item{S1, S2:}{choose from 'FECAL','ORAL','SKIN'}

\item{nOTU:}{the number of OTUs to use}

\item{pat:path}{to the otu_table.biom and 97_otus.tree}
}
\description{
Association study for American Gut project
}
