% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcov.R
\name{arcov}
\alias{arcov}
\title{Generate autoregressive cov matrix}
\usage{
arcov(r, p)
}
\arguments{
\item{r:}{correlation for neighboring variable}

\item{p:}{dimension}
}
\value{
a cov matrix
}
\description{
Generate autoregressive cov matrix
}
