% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic.cca}
\alias{bic.cca}
\title{BIC functions for CCA}
\usage{
bic.cca(n, b, r)
}
\arguments{
\item{n:}{sample size}

\item{b:}{estimated coefficients, vector}

\item{r:}{canonical correlation coefficient}

\item{g:}{a parameter in gbic and hbic}
}
\value{
bic, a scalar
}
\description{
BIC functions for CCA
}
