% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancor2.R
\name{cancor2}
\alias{cancor2}
\title{Return cca correlation}
\usage{
cancor2(x, y, ratio = 2)
}
\arguments{
\item{x, y:}{data}

\item{ratio:}{n/p, when to use identity matrix as cor}
}
\description{
Return cca correlation
}
