% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora}
\alias{carnivora}
\title{Geographical distributions of the mammalian order Carnivora}
\format{
A matrix with 41,580 rows (sites) and and 288 columns (species)
\describe{
  \item{\strong{cell value}}{absence-presence (0/1) values}
  \item{\strong{row name}}{label which can be used to retrieve the site specs from the data frame \code{carnivora_sites}}
  \item{\strong{column name}}{label which can be used to retrieve the species name from the data frame \code{carnivora_species}}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.bp26v20}
}
\usage{
carnivora
}
\description{
This common-used dataset in Ecology, gives the geographical distributions of the mammalian order Carnivora.
The data set is comprised by an incidence matrix M
(i.e. presence-absence matrix), with 41,580 non-empty sites (rows) and 288 extant terrestrial and marine species (columns).
The sites represent grid-cells rasterised at a resolution of 0.78 latitudinal degrees.
The distributional data results from rasterizing range map distributions published as part of
the Phylacine v1.2 dataset (Faurby et al. 2018), which includes
data on the geographic distributions of mammals both extant and extinct over the last 130,000 years.
The data was downloaded (last accessed on November 2019; \url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.bp26v20}) and pruned
to only include carnivores.
Data was processed in R (R Core Development Team 2014) and mapped in QGIS v2.18.16 (QGIS Development Team 2015).
}
\keyword{datasets}
