% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora_old}
\alias{carnivora_old}
\title{Carnivora Incidence in Marine Territoria (old verion)}
\format{An object of class \code{matrix} with 66960 rows and 315 columns.}
\usage{
carnivora_old
}
\description{
Carnivora Incidence in Marine Territoria (old verion)
}
\keyword{datasets}
