% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora_species}
\alias{carnivora_species}
\title{Species of the Carnivora incidence in marine territoria.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 288 rows and 2 columns.}
\usage{
carnivora_species
}
\description{
This table has two columns: species_id and species_name. The first column contains labels (column names) of the Carnivora
incidence matrix. The second column contains the names of the species.
}
\keyword{datasets}
