% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_cor.R
\name{cca.cor}
\alias{cca.cor}
\title{Re-estimate the canonical correlaiton}
\usage{
cca.cor(x, y, u, v)
}
\arguments{
\item{x, y:}{data}

\item{u, v:}{estiamted coefficients}
}
\description{
Re-estimate the canonical correlaiton
}
