% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centers.R
\name{compute_centers}
\alias{compute_centers}
\title{Compute Cluster Centers}
\usage{
compute_centers(k)
}
\arguments{
\item{k}{Integer; the number of relevant Eigenvalues}
}
\description{
Computes k cluster centers for \code{kmeans} to start with.
}
