% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_pre_process.R
\name{compute_rca}
\alias{compute_rca}
\title{Transform Bi-partite Matrix to RCA Matrix}
\usage{
compute_rca(m, binary = TRUE)
}
\arguments{
\item{m}{A data bi-partite data matrix}

\item{binary}{If TRUE binary value (0/1) of RCA is calculated}
}
\description{
Transform Bi-partite Matrix to RCA Matrix
}
