% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_symmetric.R
\name{compute_symmetric}
\alias{compute_symmetric}
\title{Compute Symmetric Matrix of a bipartite adjacency matrix}
\usage{
compute_symmetric(matrix, decomp_axis)
}
\arguments{
\item{matrix}{A numeric matrix}

\item{decomp_axis}{Character string with value 'row' or 'column'.}
}
\description{
Compute Symmetric Matrix of a bipartite adjacency matrix
}
