% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_y.R
\name{create_y}
\alias{create_y}
\title{Mysterious function}
\usage{
create_y(sub_matrix, eigenvectors, eigenvalues, k)
}
\arguments{
\item{sub_matrix}{Numeric matrix; Laplacian}

\item{eigenvectors}{Matrix; columns are eigenvectors of sub_matrix}

\item{eigenvalues}{Numeric vector}

\item{k}{Numeric; the number of eigenvalues to be considered}
}
\description{
Mysterious function
}
