% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_heuristic.R
\name{eigengap_heuristic}
\alias{eigengap_heuristic}
\title{Heuristic to Calculate the Expected Number of Clusters and Embedding Matrix}
\usage{
eigengap_heuristic(eigenvalues, eigenvectors)
}
\arguments{
\item{eigenvalues}{Numeric vector of eigenvalues}

\item{eigenvectors}{Numeric matrix containing eigenvectors (columns)}
}
\value{
A list with 3 elements
\describe{
  \item{Y}{Matrix with observations as input for \emph{kmeans}}
  \item{k}{The number of expected centers}
}
}
\description{
Given a spectrum (decreasingly sorted Eigenvalues) and the corresponding Eigenvalues \emph{eigengap_heuristic}
computes the number of centers and the input matrix (\emph{embedding} for \emph{kmeans}.
}
