% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_cca.R
\name{elnet.cca}
\alias{elnet.cca}
\title{Elastic net cca}
\usage{
elnet.cca(x, y, a.u, a.v, l.u, l.v)
}
\arguments{
\item{a.u, a.v:}{The elasticnet mixing parameter}

\item{l.u, l.v:}{lambda*penalty}
}
\value{
estimated coefficients, optimal prameters
}
\description{
Elastic net cca
}
