% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_simu.R
\name{elnet.cca.simu}
\alias{elnet.cca.simu}
\title{Simulation for elastic net cca}
\usage{
elnet.cca.simu(niter, u, v, n = 50, data.type = "normal", sigma_z = 1,
  sigma_x = NULL, sigma_y = NULL, theta = 0.01, n_x = 1000,
  n_y = 1000, lambda_z = 10, cv.method = "bic", alpha.u = c(0.6, 0.8,
  1), alpha.v = c(0.6, 0.8, 1), K = 5, verbose = F)
}
\arguments{
\item{alpha.u, alpha.v:}{The elasticnet mixing parameter, between 0 and 1.}
}
\description{
Simulation for elastic net cca
}
