% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exports}
\alias{exports}
\title{Exports}
\format{
Matrix of 234 rows (countries) and 1239 columns (products)
\describe{
  \item{\strong{cell value}}{Value of the export in US dollars in 2016.}
  \item{\strong{row name}}{label of the exporting country which can be used to retrieve country data from \code{exports_countries}.}
  \item{\strong{column name}}{label of the exported product which can be used to retrieve product data from \code{exports_products}.}
}
}
\source{
\url{http://atlas.cid.harvard.edu/about-data}
}
\usage{
exports
}
\description{
An adjacency matrix of 234 countries (rows) exporting 1239 products (columns).
The each cell contains the value of the export in US dollars in 2016. See vignette X an CA of this data with SCCA.
See \strong{details} for description of the data.
}
\details{
The matrix is derived from trade data originally collected by the United Nations Statistical Division (COMTRADE) that was cleaned,
processed and made publicly available by the Growth Growth Lab at Harvard University (\url{http://atlas.cid.harvard.edu/about-data}).
}
\keyword{datasets}
