% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exports_countries}
\alias{exports_countries}
\title{Country names of exporting countries and their GDPs}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 234 rows and 4 columns.
}
\usage{
exports_countries
}
\description{
See: \link[=exports]{exports}
}
\keyword{datasets}
