% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exports_products}
\alias{exports_products}
\title{Products of the Exports data}
\format{
Data frame of 1239 rows (products) and 2 columns
\describe{
  \item{\strong{label}}{Label of the exported product used in matrix \code{exports} as column name.}
  \item{\strong{product_name}}{Name of the product}
}
}
\usage{
exports_products
}
\description{
Products of the Exports data
}
\keyword{datasets}
