% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finals.R
\name{finals}
\alias{finals}
\title{Summary of simulation results}
\usage{
finals(object)
}
\arguments{
\item{object:}{an object from simulation function}
}
\value{
TPR,FPR,F1,Cor
}
\description{
Summary of simulation results
}
