% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count _data.R
\name{GenStrucData_Count_latent}
\alias{GenStrucData_Count_latent}
\title{Generate count data}
\usage{
GenStrucData_Count_latent(u, v, n = 50, lambda_z = 10, theta = 0.01,
  n_x = 1000, n_y = 1000)
}
\arguments{
\item{lambda_z:}{poisson parameter of the latent variable z}

\item{theta:}{oversidpersion parameter in DM model}

\item{n_y:}{parameter n in DM model}

\item{n_x:}{parameter n in DM model}

\item{u:}{caonical coefficients for x}

\item{v:}{canonical coefficients for y}

\item{n:}{sample size}
}
\description{
Generate count data using latent variable model
}
