% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_data.R
\name{GenStrucData_Normal}
\alias{GenStrucData_Normal}
\title{Generate normal data using latent variable model}
\usage{
GenStrucData_Normal(u, v, n = 50, sigma_z = 1, sigma_x = NULL,
  sigma_y = NULL)
}
\arguments{
\item{u:}{caonical coefficients for x}

\item{v:}{canonical coefficients for y}

\item{n:}{sample size}

\item{sigma_z:}{sd of the latent variable z, scalar}

\item{sigma_x:}{covariance matrix of x, default to be identity}

\item{sigma_y:}{covariance matrix of y, default to be identity}
}
\value{
a list containing data mattrix X and Y
}
\description{
Generate normal data using latent variable model
}
