% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_data_misspecification.R
\name{GenStrucData_Normal_mis}
\alias{GenStrucData_Normal_mis}
\title{Generate normal data, misspecification}
\usage{
GenStrucData_Normal_mis(u, v, n = 50, sigma_z = 1, sigma_x = NULL,
  sigma_y = NULL)
}
\description{
Generate normal data, misspecification
}
