% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_data_orthogonal.R
\name{GenStrucData_Normal_oth}
\alias{GenStrucData_Normal_oth}
\title{Generate normal data, orthogonal direction}
\usage{
GenStrucData_Normal_oth(u, v, n = 50, sigma_z = 1, sigma_x = NULL,
  sigma_y = NULL)
}
\value{
data x and y
}
\description{
Generate normal data, orthogonal direction
}
