% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_matrix2.R
\name{GenStrucMat}
\alias{GenStrucMat}
\title{Function to generate structure matrix using ape tree object}
\usage{
GenStrucMat(x, tree, p_c = 1, dis = F, p_d = 1, h = 0.1, wcorr = F)
}
\arguments{
\item{x:}{data matrix}

\item{tree:}{tree object}

\item{p_c:}{|r_ij|^p_c}

\item{p_d:}{d^p_d}

\item{dis:}{whether distance weighted;}

\item{wcorr:}{whether using correlation weighted}

\item{h:}{controls the complexity of structural penalty. Larger h, more complex. h should between 0 and 1.}
}
\description{
This function takes tree object from ape package as input.
}
