% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_validation_test.R
\name{get_clustering}
\alias{get_clustering}
\title{Get Clustering from an SCCA run}
\usage{
get_clustering(scca)
}
\arguments{
\item{scca}{An SCCA analysis tree; output from a run of scca_compute on a dataset}
}
\value{
A tibble with two columns: label and cluster. Each row represents a case identified by
 a label and it's cluster.
}
\description{
Get Clustering from an SCCA run
}
\details{
get_clustering is a wrapper around get_clustering_recursive
}
