% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_overlap_test.R
\name{get_clustering_recursive}
\alias{get_clustering_recursive}
\title{Get Cluster Recursively}
\usage{
get_clustering_recursive(scca, id = 0, leaves_only = TRUE)
}
\arguments{
\item{scca}{An SCCA tree or sub-tree}

\item{id}{Integer; the number of the last found node. Equals 0 when no cluster has been found yet.}

\item{leaves_only}{Boolean, default is TRUE. Not implemented. Do not use!}
}
\description{
Descends recursively the analysis trees and collects the final clusters (leaves) in one dataframe (tibble)
}
