% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_overlap_test.R
\name{get_clusters_recursive}
\alias{get_clusters_recursive}
\title{Get Clusters}
\usage{
get_clusters_recursive(scca, cluster_path = NULL, leaves_only = TRUE)
}
\arguments{
\item{scca}{An SCCA tree}

\item{cluster_path}{Sequence of child numbers (seperated by '.') visited from top node to this node.}

\item{leaves_only}{Boolean, default is TRUE. Use default.}
}
\description{
Descends recursively the analysis tree and collects the final clusters at the leaves.
}
