% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{gscca.permute}
\alias{gscca.permute}
\title{Permutation Test}
\usage{
gscca.permute(nperm = 100, x, y, edgex, edgey, maxsteps = 20, plain = T,
  cv.method = "bic", gamma.u, gamma.v, lambda.u, lambda.v, Sx = NULL,
  Sy = NULL, ccor = F, thresh = 0.5, seed = NULL)
}
\arguments{
\item{nperm:}{the number of permutation test}
}
\value{
p.corr,p.corr.re: p values

corrs,corrs.re: all correlations from permutation

corr0,corr.re0: optimal correlation
}
\description{
Perform permutaiton test of gscca
}
