% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscca.R
\name{gscca}
\alias{gscca}
\title{Implement gscca model}
\usage{
gscca(x, y, edgex, edgey, maxsteps = 2000, minlam.u = NULL,
  minlam.v = NULL, plain = T, Sx = NULL, Sy = NULL)
}
\arguments{
\item{minlam:}{the minimum lambda at which path algo should stop.}
}
\value{
canonical correlation coefficients
}
\description{
Implement gscca model
}
