% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscca_roc.R
\name{gscca.roc}
\alias{gscca.roc}
\title{ROC curve, gscca}
\usage{
gscca.roc(x, y, u, v, edgex, edgey, maxsteps = 20, plain = T, gamma.u,
  gamma.v, lambda.u = NULL, lambda.v = NULL, Sx = NULL, Sy = NULL)
}
\value{
a list containing TPR,FPR
}
\description{
ROC curve, gscca
}
