% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscca_simu.R
\name{gscca.simu.normal}
\alias{gscca.simu.normal}
\title{Simulation for gscca}
\usage{
gscca.simu.normal(niter, edges, u, v, n = 50, data.type = "normal",
  sigma_z = 1, sigma_x = NULL, sigma_y = NULL, theta = 0.01,
  lambda_z = 10, n_x = 1000, n_y = 1000, cv.method = "bic",
  gamma.u = NULL, gamma.v = NULL, lambda.u = NULL, lambda.v = NULL,
  maxsteps = 20, plain = T, Sx = NULL, Sy = NULL, verbose = F,
  K = 5)
}
\arguments{
\item{gamma.u, gamma.v:}{ratio of l1 peanlty and fused penalty}

\item{lambda.u, lambda.v:}{fused penalty. lambda*fused penalty+gamma*lambda*l1 penalty.}

\item{Sx, Sy:}{structure matrix for Sx,Sy.}

\item{maxsteps:}{the number of lambdas to try, default to be 20. Largest 20 values from solution path}
}
\description{
Simulation for gscca
}
