% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscca_simu_tree.R
\name{gscca.simu}
\alias{gscca.simu}
\title{Simulation for gscca}
\usage{
gscca.simu(niter, u, v, n = 50, data.type = "normal", sigma_z = 1,
  sigma_x = NULL, sigma_y = NULL, theta = 0.01, lambda_z = 10,
  n_x = 1000, n_y = 1000, cv.method = "bic", gamma.u = NULL,
  gamma.v = NULL, edgex = NULL, edgey = NULL, Sx = NULL, Sy = NULL,
  p_c = 1, dis = F, p_d = 1, h = 1, wcorr = T, thresh = 0.5,
  lambda.u = NULL, lambda.v = NULL, maxsteps = 20, plain = T,
  verbose = F, K = 5, ccor = F, seed = 123)
}
\arguments{
\item{gamma.u, gamma.v:}{ratio of l1 peanlty and fused penalty}

\item{Sx, Sy:}{structure matrix for Sx,Sy.}

\item{lambda.u, lambda.v:}{fused penalty. lambda*fused penalty+gamma*lambda*l1 penalty.}

\item{maxsteps:}{the number of lambdas to try, default to be 20. Largest 20 values from solution path}

\item{ccor:}{whether re-estiamte the coeeficient when doing BIC}
}
\description{
This function generate tree from data each time instead of using same tree for all simulations
}
