% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make)edge.R
\name{Make_edge}
\alias{Make_edge}
\title{Generate edges given tips}
\usage{
Make_edge(tip)
}
\arguments{
\item{tip:}{tips, a matrix, the number of column is 2.}
}
\value{
edges for generating structural matrix
}
\description{
Generate edges given tips
}
