% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a random variable}
\usage{
normalize(x)
}
\arguments{
\item{x:}{input data matrix}
}
\value{
normalized column, matrix
}
\description{
Normalize a random variable
}
