% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_overlap_test.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot the overlap between two clusterings of the same dataset and category}
\usage{
plot_overlap(overlap_xy)
}
\arguments{
\item{overlap_xy}{A tibble with overlaps between pairs of clusters}
}
\description{
Plot the overlap between two clusterings of the same dataset and category
}
\note{
Do not use; contains bugs
}
