% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_simu.R
\name{roc.simu}
\alias{roc.simu}
\title{Simulate the ROC}
\usage{
roc.simu(nroc = 1, u, v, n = 50, data.type = "normal", sigma_z = 1,
  sigma_x = NULL, sigma_y = NULL, theta = 0.01, n_x = 1000,
  n_y = 1000, lambda_z = 10, maxstep_s0 = NULL, penaltyx = seq(0, 1,
  length.out = 50), penaltyy = seq(0, 1, length.out = 50), alpha.u = seq(0,
  1, length.out = 10), alpha.v = seq(0, 1, length.out = 10), nlambda = 20,
  edges, maxstep_gs = 30, gamma.u = seq(0, 5, length.out = 10),
  gamma.v = seq(0, 5, length.out = 10))
}
\arguments{
\item{maxstep_s0, maxstep_gs:}{paramters for s0cca and gscca}

\item{penaltyx, penaltyy:}{for s1cca, vectors}

\item{alpha.u, alpha.v, nlambda:}{for enscca, vector, vector, scalar.}

\item{gamma.u, gamma.v:}{paramters for gscca, vectors.}

\item{nroc:}{the number of iterations}

\item{data.type:}{type of data to generate, choose from 'normal','normal_oth','normal_oth'...}
}
\description{
A function to run the simulation for plotting ROC
}
