% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s0cca_final.R
\name{s0cca.final}
\alias{s0cca.final}
\title{Final s0cca model calculation}
\usage{
s0cca.final(x, y, lambda, covI = F)
}
\arguments{
\item{x, y:}{data matrix}

\item{lambda:}{thresholding parameter}

\item{covI:}{If ture, using identity matrix for cov(x) and cov(y).}
}
\value{
estimated coefficients and correlation
}
\description{
Final s0cca model calculation
}
