% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s0cca.R
\name{s0cca}
\alias{s0cca}
\title{Conduct s0cca}
\usage{
s0cca(x, y, ncluster = NULL, cluster.type = "no", maxsteps = 20,
  cv.method = "cv", K = 5, covI = T)
}
\arguments{
\item{x, y:}{data matrix}

\item{ncluster:}{the number of cluster when clustering  the lambda. If NULL, using all lambdas.}

\item{cluster.type:}{how to cluster lambdas, choose from 'no'(do not do the clustering), 'seq','quantile','kmeans'.}

\item{maxsteps:}{how many lambdas to try, default to be 20.}

\item{cv.method:}{choose from 'cv','bic','ebic','hbic','gic'.}
}
\value{
optimal canonical coeffcients, canonical correlation, optimal lambda.
}
\description{
Conduct s0cca
}
