% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s0cca_roc.R
\name{s0cca.roc}
\alias{s0cca.roc}
\title{ROC curve, s0cca}
\usage{
s0cca.roc(x, y, u, v, ncluster = NULL, cluster.type = "no", maxsteps = 20)
}
\value{
a list containing TPR,FPR
}
\description{
ROC curve, s0cca
}
