% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s0cca_cv.R
\name{s0cca.thresh}
\alias{s0cca.thresh}
\title{CV function for s0cca}
\usage{
s0cca.thresh(x, y, u, v, covI = T)
}
\arguments{
\item{x, y:}{data}

\item{u, v:}{vector indcating the sparseness of coefficients}
}
\value{
estimated coefficients and correlation
}
\description{
For each fold of CV, return the estimated coefficients and correlation
}
