% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s0cca_simu.R
\name{s0cca_simu}
\alias{s0cca_simu}
\title{Simulation function for s0cca}
\usage{
s0cca_simu(niter, u, v, n = 50, data.type = "normal", sigma_z = 1,
  sigma_x = NULL, sigma_y = NULL, theta = 0.01, lambda_z = 10,
  n_x = 1000, n_y = 1000, ncluster = NULL, cluster.type = "no",
  maxsteps = 20, cv.method = "cv", verbose = F)
}
\arguments{
\item{niter:}{the number of simulation iterations}

\item{data.type:}{'normal', 'count_normal','count_latent'}
}
\value{
TPR,FPR,F1 score
}
\description{
Simulation function for s0cca
}
