% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1cca_simu.R
\name{s1cca_simu}
\alias{s1cca_simu}
\title{Simulation for s1cca}
\usage{
s1cca_simu(niter, u, v, n = 50, data.type = "normal", sigma_z = 1,
  sigma_x = NULL, sigma_y = NULL, theta = 0.01, n_x = 1000,
  n_y = 1000, lambda_z = 10, verbose = F)
}
\description{
Simulation for s1cca
}
