% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{scca}
\alias{scca}
\title{Spectral Clustering Correpondence Analysis.}
\usage{
scca(M)
}
\arguments{
\item{M}{A absence-presence matrix ....}
}
\value{
Returns a list with .....
}
\description{
\code{scca} performs a spectral analysis ..,.,.
}
\examples{
\dontrun{
data('carnivora', package = 'SCCAR')
scca(M = carnivora$M)
}
}
