% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute.R
\name{scca_compute}
\alias{scca_compute}
\title{Spectral Clustering Correpondence Analysis.}
\usage{
scca_compute(m)
}
\arguments{
\item{m}{A matrix representing a bi-partite graph. The matrix must have row names and column names}
}
\value{
A recursive which describes the hierarchical clustering process.
Each node in the tree represents a stage in the analysis. The subnodes  :
\describe{
  \item{level}{The level in the tree.}
  \item{child}{Numbering of the siblings. No order intended}
  \item{spectrum}{Vector of the Eigen values found at this node. The Eigenvalues are
  sorted on explained variance in descending order.}
  \item{labels}{The labels of the cluster in this node}
  \item{eigen_vec_1}{The first Eigenvector found in the cluster of this node}
  \item{eigen_vec_2}{The second}
}
}
\description{
The function \code{scca_compute} performs a spectral clustering correspondence analysis on a given
bi-partite graph. It is a form a hierarchical clustering. On every level the sub-clustering is guided by
a spectral analysis to estimate the number of sub-clusters (the k of kmeans). See .....
}
\examples{
\dontrun{
data('carnivora', package = 'sccar')
scca_compute(M = carnivora$M)
}
}
