% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_stability_test.R
\name{scca_compute_dist}
\alias{scca_compute_dist}
\title{SCCA compute and store distance matrix}
\usage{
scca_compute_dist(m, filename = NULL, method = "euclidian")
}
\arguments{
\item{m}{Incidence or adjacency matrix}

\item{filename}{Name of the file in which to store the distance matrix. If NULL (default) the
distance will not be stored, but only returned}

\item{method}{See ?stats::dist}
}
\value{
The distance matrix as an dist object
}
\description{
SCCA compute and store distance matrix
}
