% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree1.R
\name{scca_compute_tree1}
\alias{scca_compute_tree1}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree1(labels, child, m, level, decomp_axis)
}
\arguments{
\item{labels}{Character vector; The labels defining the rows/columns of the submatrix}

\item{child}{Numeric. Child number among its siblings (= same parent) in the tree}

\item{m}{Numeric matrix; Laplacian}

\item{level}{Integer; the depth of the node in the tree}

\item{decomp_axis}{Vector; should the rows or columns be subsetted}
}
\description{
Build recursively a scca tree
}
\note{
Expects an matrix M accessible from within its environment
}
