% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_overlap_test.R
\name{scca_get_clustering}
\alias{scca_get_clustering}
\title{Get clustering from an output of an SCCA run}
\usage{
scca_get_clustering(scca)
}
\arguments{
\item{scca}{An SCCA analysis tree; output from a run of scca_compute on a dataset}
}
\value{
A tibble with two columns: label and cluster. Each row represents a case identified by
 a label and it's cluster.

Returns a tibble. Each row contains a label of a case and the id of the cluster it's assigned to.
}
\description{
Get clustering from an output of an SCCA run
}
\details{
get_clustering is a wrapper around get_clustering_recursive
}
