% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_print.R
\name{scca_get_node}
\alias{scca_get_node}
\title{Retrieve Attributes of a Node in an SCCA tree}
\usage{
scca_get_node(scca, node)
}
\arguments{
\item{scca}{An SCCA tree}

\item{node}{The number of the node to retrieve the attributes from. The supported attributes are
currently \emph{spectrum} and \emph{labels}}
}
\value{
List with two elements:
\describe{
  \item{labels}{The labels (rownames) of the (sub-)cluster in this node}
  \item{spectrum}{The Eigenvalues of the data matrix}
}
}
\description{
Retrieve Attributes of a Node in an SCCA tree
}
\details{
Returns NULL if \emph{node} doesn't exists.
}
