% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_node.R
\name{scca_node}
\alias{scca_node}
\title{Build recursively a SCCA tree}
\usage{
scca_node(M, labels, level, axis)
}
\arguments{
\item{M}{Numeric matrix; Laplacian}

\item{labels}{Character vector; The labels defining the rows/columns of the submatrix}

\item{level}{Integer; the depth of the node in the tree}

\item{axis}{Vector; should the rows or columns be subsetted}
}
\description{
Build recursively a SCCA tree
}
\note{
Expects an matrix M accessible from within its environment
}
