% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_overlap_test.R
\name{scca_overlap_test}
\alias{scca_overlap_test}
\title{SCCA Overlap Test}
\usage{
scca_overlap_test(x, y, plot = FALSE)
}
\arguments{
\item{x}{An output tree of an SCCA run}

\item{y}{An output of another SCCA run on the same dataset/category as in x}

\item{plot}{Boolean; plot an bi-partite, overlap graph between clusters of x and y. Default is FALSE}
}
\value{
A list with 3 elements:
\describe{
  \item{avg_overlap.x}{The chance that a pair of cases which are in the same cluster of x
  are also in the same cluster of y. The weighted average of overlaps of all clusters in x with clusters in y }
  \item{avg_overlap.y}{The weighted average of overlaps of all clusters in y }
  \item{overlap.xy}{Tibble. Overlap between each (x,y) pair of clusters}
}
}
\description{
The function \code{scca_overlap_test} compares two SCCA clusterings of the same dataset and category to
establish their ovelap. It calculates an average proportion of overlap between the two clusterings.
}
\examples{
\dontrun{
data('carnivora', package = 'sccar')
sc1 <- scca_compute(carnivora)
sc2 <- scca_compute(carnivora)
scca_overlap_test(x = sc1, y = sc2)
}
}
