% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_print.R
\name{scca_print}
\alias{scca_print}
\title{Print SCCA Analysis}
\usage{
scca_print(scca, ...)
}
\arguments{
\item{scca}{An output of scca_compute}

\item{...}{The attributes to be printed: 'depth', 'k', 'node_type' and 'nlabs'}
}
\description{
Prints the hierarchical analysis to screen in character-based, human-readable fashion
}
\details{
The numbers in the tree are the child numbers. \code{depth} is the level in the tree and \code{nlabs} the number of
  of labels in the (sub-)cluster. Each attribute can be ommited. If no attribute is given, then only the hierarchy is printed.
}
\examples{
  \dontrun{
     s <- scca_compute(t(carnivora))
     scca_print(scca = s, 'k', 'nlabs')
}
}
