% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_py_compare.R
\name{scca_py_overlap_test}
\alias{scca_py_overlap_test}
\title{Overlap Test between an SSCA clustering and a 'Python clustering'}
\usage{
scca_py_overlap_test(scca, py_output = NULL, plot = FALSE)
}
\arguments{
\item{scca}{An SSCA tree.}

\item{py_output}{The path from the working directory to the directory with the Python output files.
If NULL (default), then the working directory is taken as path.}

\item{plot}{If TRUE an overlap graph is plotted; default is FALSE.}
}
\description{
The \strong{sccarr package} replaces an implementation of SCCA in Python.
\emph{scca_py_overlap_test} compares clusterings from both implementations by calculating
the average proportion of overlap.
For more info see \link{scca_overlap_test}.
In due time this function will be phased out.
}
\details{
For more info see \code{\link{scca_overlap_test}}.

  In due time this function will be phased out.
}
