% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_validity_test.R
\name{scca_silhouette_test}
\alias{scca_silhouette_test}
\title{Silhouette Width of each Cluster in an SCCA Clustering}
\usage{
scca_silhouette_test(scca, dist, plot = TRUE)
}
\arguments{
\item{scca}{An SCCA clustering tree}

\item{dist}{A \emph{dist} object; a distance matrix of the data used for producing \emph{scca}}

\item{plot}{Boolean; if TRUE (default) plots the silhouette width of every obeservation}
}
\description{
Silhouette Width of each Cluster in an SCCA Clustering
}
