% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_stability_test.R
\name{scca_stability_test}
\alias{scca_stability_test}
\title{SCCA Stability Test}
\usage{
scca_stability_test(m, del_columns = NULL, dist_m = NULL)
}
\arguments{
\item{m}{Matrix representing an incidence (or bipartite) network}

\item{del_columns}{Integer vector, the index of the columns to be deleted (one for one) when testing
the stability. When omitted, all the columns are deleted (one for one)}

\item{dist_m}{Object of class dist; the distance matrix that corresponds with the incidence matrix m. If NULL (default),
scca_stability_test calculates the distance matrix}
}
\description{
Tests the stability of the clustering method.
}
\details{
Performs an SCCCA stability test on a clustering of a dataset.
First, a clustering on the complete dataset is done.
This is the base clustering.
Then, one for one, each variable in the dataset is removed and a new
clustering is calculated.
Each new clustering is compared with the base clustering.
The measures are APN, AD, ADM and FOM.
The returned values are averages of the outcomes of the measure per removed variable.
If the user provides a list of columns, only those columns will be taken into account.
}
