% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_validity_test.R
\name{scca_validity_test}
\alias{scca_validity_test}
\title{SCCA Validity Test}
\usage{
scca_validity_test(scca, dist)
}
\arguments{
\item{scca}{List; an SCCA clustering tree}

\item{dist}{A \emph{dist} object; the distance matrix of the data used for producing \emph{scca}}
}
\value{
A list with the three measures
}
\description{
Tests the validity of a clustering produced by \emph{scca_compute}.
}
\details{
Computes the three internal validity measures: \strong{Connectivity}, \strong{Silhouette Width} and \strong{Dunn Index}. See
\href{https://cran.r-project.org/web/packages/clValid/vignettes/clValid.pdf}{clValid, an R package for cluster validation} for more
information
}
