% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scca.R
\name{scca_write_analysis}
\alias{scca_write_analysis}
\title{Write  a SCCA to File}
\usage{
scca_write_analysis(analysis_tree, vec_name, spec_name)
}
\arguments{
\item{analysis_tree}{Output tree of an SCCA}

\item{vec_name}{Prefix of the names for files with the Eigenvectors}

\item{spec_name}{Prefix of the filenames with the spectra}
}
\value{
TRUE
}
\description{
Writes the results of an analysis to .csv files
}
