% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_write_clustering.R
\name{scca_write_clustering}
\alias{scca_write_clustering}
\title{Write an Output of SCCA to Files}
\usage{
scca_write_clustering(
  scca_tree,
  vec_name = "V_",
  spec_name = "S_",
  leaves_only = TRUE
)
}
\arguments{
\item{scca_tree}{list; the tree resulting from a call to 'scca_compute'}

\item{vec_name}{Prefix  (string) of the names of files with the labels and the Eigenvectors}

\item{spec_name}{Prefix of the names of he files with the spectra}

\item{leaves_only}{If TRUE (default) only the final clusters (leaf nodes) are written}
}
\value{
TRUE
}
\description{
Writes the results of an SCCA to .csv files. For every tree node two files are written:
}
\details{
\describe{
  \item{Eigenvectors}{The first column contains labels of the subcluster which is analyzed at the node,
and columns 2 to 4 are the 3 most prominent Eigenvectors}
  \item{Spectrum}{The first column contains the spectrum of the subcluster and the second column the explained variance}
}

The filename contains the path from the top to the tree node.
}
