% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccar.R
\docType{package}
\name{sccar}
\alias{sccar}
\title{sccar: An R package for Spectral Clustering Correspondence Analysis.}
\description{
R functions for ....
}
\section{sccar functions}{

The exported functions ...
}

