% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_matrix.R
\name{StrucMat}
\alias{StrucMat}
\title{Generate Structure Matrix in gscca}
\usage{
StrucMat(edges, x, plain = T)
}
\arguments{
\item{edges:}{the edges of graph structure, see peacakge igraph}

\item{x:}{data matrix}

\item{plain:}{whether using weighted penalty. If True, the penalty is $\Sigma |\beta_i-\beta_j|$. If false, the peanlty is $\Sigma |r_ij||\beta_i-sign(r_ij)\beta_j|$.}
}
\value{
structure matrix, D
}
\description{
Generate Structure Matrix in gscca
}
