% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPRFPR.R
\name{TruePR}
\alias{TruePR}
\title{TPR and FPR}
\usage{
TruePR(label, pred)
}
\arguments{
\item{label:}{true label, vector, elements are either 0 or 1}

\item{pred:}{predicted value, vector, elements are either 0 or 1}
}
\value{
TPR, FPR
}
\description{
TPR and FPR
}
