% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_heuristic.R
\name{apply_heuristic}
\alias{apply_heuristic}
\title{Apply scca heuristic}
\usage{
apply_heuristic(eigenvalues)
}
\arguments{
\item{eigenvalues}{Numeric vector of eigenvalues}
}
\value{
The number of weakly connected components (possibly clusters)
}
\description{
The function \code{apply_heuristic} implements heuristic for SCCA
-> Returns index and vectors of # weakly connected components
-> Calculate gaps, check for disconnect components, and otherwise for the largest gap
}
