% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora}
\alias{carnivora}
\title{Carnivora incidence in marine territoria.}
\format{A list with to elements
\describe{
  \item{M}{matrix with }
  \item{E}{dataframe with }
  ...
}}
\source{
\url{http://www.somewhere}
}
\usage{
carnivora
}
\description{
The incidences of certain species (carnivora) at marine sites are given by a bipartite adjacency matrix.
The rows are the sites and the columns are the species.
The dataset also has a table with for every site the lon/lat (which projection?)
}
\keyword{datasets}
