% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora_sites}
\alias{carnivora_sites}
\title{Site coordinates of the Carnivora dataset}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41580 rows and 3 columns.}
\usage{
carnivora_sites
}
\description{
The Carnivora dataset also has a table with for every site the row id in the incidence matrix. The second and third columns give
the coordinates (lon/lat) of the site. The Behrman cylindrical equal area projection is used with
a resolution of 96.5 km by 96.5 km at 30° North and 30° South.
The full projection description following R-notation is
“+proj=cea +lon_0=0+lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84+towgs84=0,0,0”.
Only the area between 90°N and 60°S is included.
}
\keyword{datasets}
