% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora_species}
\alias{carnivora_species}
\title{Species of the Carnivora Incidence in Marine Territoria.}
\format{
A data frame with 288 rows (species)
\describe{
  \item{species_id}{Species labels used in matrix \code{carnivora} as column names}
  \item{species_name}{The actual names of the species}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.bp26v20}
}
\usage{
carnivora_species
}
\description{
This table has two columns: species_id and species_name. The first column contains labels (column names) of the Carnivora
incidence matrix. The second column contains the names of the species.
}
\keyword{datasets}
