% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_pre_process.R
\name{compute_rca}
\alias{compute_rca}
\title{Compute RCA Matrix}
\usage{
compute_rca(m, binary = TRUE)
}
\arguments{
\item{m}{A data matrix representing a bi-partite network}

\item{binary}{If TRUE (default) binary RCA values are calculated}
}
\description{
Computes Revealed Comparative Advantage (a.k.a. Location Quotient) of a bi-partite network (e.g. country <-> exported products)
}
\details{
See: \href{https://en.wikipedia.org/wiki/Revealed_comparative_advantage}{Wikipedia}.
}
