% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_pre_process.R
\name{compute_rca}
\alias{compute_rca}
\title{Compute RCA Matrix}
\usage{
compute_rca(m, binary = TRUE)
}
\arguments{
\item{m}{A data matrix representing a bi-partite graph}

\item{binary}{If TRUE (default) binary version is calculated}
}
\description{
Computes Revealed Comparative Advantage (a.k.a. Location Quotient) from a country/products matrix
(like the exports dataset)
If RCA > 1 the product is 'over-represented' in a country.
A binary version has values 1 if RCA > 1 and 0 otherwise.
See: \href{https://en.wikipedia.org/wiki/Revealed_comparative_advantage}{Wikipedia}.
}
