% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_pre_process.R
\name{compute_rca}
\alias{compute_rca}
\title{Compute RCA Matrix}
\usage{
compute_rca(m, binary = TRUE)
}
\arguments{
\item{m}{A data matrix}

\item{binary}{If TRUE (default) the function returns a binary matrix with entry 1 if RCA>1 and 0 otherwise}
}
\description{
Computes Revealed Comparative Advantage (a.k.a. Location Quotient) of a an incidence matrix (e.g. country <-> exported products)
}
\details{
See: \href{https://en.wikipedia.org/wiki/Revealed_comparative_advantage}{Wikipedia}.
}
