% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_symmetric.R
\name{compute_symmetric}
\alias{compute_symmetric}
\title{Compute Eigenvalues and Eigenvectors of Large and Sparse Incidence Matrix}
\usage{
compute_symmetric(matrix, decomp_axis)
}
\arguments{
\item{matrix}{An incidence matrix. The rows are the observations and
the columns are the}

\item{decomp_axis}{Character string with value 'row' or 'column' indicating the axis along
which the decomposition will initially take place. At the end the decomposition will be transformed
to the row axis (= observations)}
}
\description{
See <<link>>
}
