% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_y.R
\name{create_y}
\alias{create_y}
\title{Mysterious function}
\usage{
create_y(sub_matrix, eigenvectors, eigenvalues, k)
}
\arguments{
\item{sub_matrix}{Numeric matrix of n * n}

\item{eigenvectors}{Numeric matrix of n * m. Columns are eigenvectors of sub_matrix. Only the first k columns
are used}

\item{eigenvalues}{Numeric vector of at least k largest eigenvalues of sub_matrix}

\item{k}{Numeric; the number of eigenvalues to be considered}
}
\description{
Mysterious function
}
