% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Y.R
\name{create_Y}
\alias{create_Y}
\title{Mysterious function}
\usage{
create_Y(subM, Eigenvecs, Eigenvals, k)
}
\arguments{
\item{subM}{Numeric matrix; Laplacian}

\item{Eigenvecs}{Matrix; columns are eigenvectors of subM}

\item{Eigenvals}{Numeric vector}

\item{k}{Numeric; the number of eigenvalues to be considered}
}
\description{
Mysterious function
}
