% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_symmetric.R
\name{decomp_symmetric}
\alias{decomp_symmetric}
\title{Decompose Large and Sparse Matrix}
\usage{
decomp_symmetric(matrix, max_eigenvalues, decomp = "svd")
}
\arguments{
\item{matrix}{An incidence matrix. The rows are the observations and
the columns are the variables}

\item{max_eigenvalues}{Max. number of eigenvalues to compute. Default is 25.}

\item{decomp}{The decomposition to use: 'svd' (default) or 'svds'.}
}
\description{
See <<link>>
}
\details{
If both dimensions of data matrix are greater than max_eigenvalues, then the number of computed eigenvalues is restricted
to max_eigenvalues else the shortest dimension is choosen.
}
