% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_symmetric.R
\name{decomp_symmetric}
\alias{decomp_symmetric}
\title{Decompose Large and Sparse Matrix}
\usage{
decomp_symmetric(matrix, max_eigenvalues, decomp = "svd")
}
\arguments{
\item{matrix}{Incidence matrix (e.g. species - location), which can be interpreted as the  bi-adjacency matrix of a bipartite network. The rows are taken to be the unit of analysis.}

\item{max_eigenvalues}{Max. number of eigenvalues to compute. Default is 25.}

\item{decomp}{The decomposition to use: \strong{svd} (default) or \strong{svds}.
The later only computes the k leading singular values and vectors of an rectangular matrix}
}
\description{
The function computes weighted singular values and correponding singular vectors
}
\details{
If both dimensions of data matrix are greater than max_eigenvalues, then the number of computed eigenvalues is restricted
to max_eigenvalues else the shortest dimension is chosen.

\strong{svds} uses the rARPACK implementation of the singular value decomposition for efficient approximation of singular decomposition for large sparse matrices.
}
