% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_heuristic.R
\name{eigengap_heuristic}
\alias{eigengap_heuristic}
\alias{trace_heuristic}
\title{Heuristics to Calculate the Expected Number of Clusters and Embedding Matrix}
\usage{
eigengap_heuristic(eigenvalues, eigenvectors)

trace_heuristic(eigenvalues, eigenvectors)
}
\arguments{
\item{eigenvalues}{Numeric vector of eigenvalues}

\item{eigenvectors}{Numeric matrix containing eigenvectors (columns)}
}
\value{
A list with 3 elements
\describe{
  \item{Y}{The embedding matrix}
  \item{min.nc}{The minimum number of cluster centers}
  \item{max.nc}{The maximum number of cluster centers}
}


}
\description{
Given a spectrum (decreasingly sorted Eigenvalues) and the corresponding Eigenvalues the heuristics compute the number of centers and
input matrix for kmeans.
}
