% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exports}
\alias{exports}
\title{Exports data.}
\format{A list with to elements
\describe{
  \item{M}{Matrix with export values of products (columns) by countries (rows)  }
  \item{country_name}{one column datatable with country names}
  ...
}}
\source{
\url{http://atlas.cid.harvard.edu/about-data}
}
\usage{
exports
}
\description{
The exports data set consists of a matrix of 234 rows that represent countries and 1239 columns that represent products.
The entries give the value of exports in 2016 in US $.
The product names are given by the 4-digit 'Harmonized System' classification.
}
\details{
The matrix is derived from trade data originally collected by the United Nations Statistical Division (COMTRADE) that was cleaned,
processed and made publicly available by the Growth Growth Lab at Harvard University.
For detailed we refer to http://atlas.cid.harvard.edu/about-data.

The Growth Lab at Harvard University, 2019, "country_partner_hsproduct4digit_year_2016.tab",
International Trade Data (HS, 92), https://doi.org/10.7910/DVN/T4CHWJ/T3KWV0, Harvard Dataverse
}
\keyword{datasets}
