% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exports_products}
\alias{exports_products}
\title{Exported products}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1239 rows and 2 columns.
}
\usage{
exports_products
}
\description{
See: \link[=exports]{exports}
}
\keyword{datasets}
