% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_py_compare.R
\name{get_py_clustering}
\alias{get_py_clustering}
\title{Get the Clustering Produced by Python Script}
\usage{
get_py_clustering(path = NULL)
}
\arguments{
\item{path}{The path from the working directory to the directory containing the Python output}
}
\value{
A tibble with 3 variables:
\describe{
  \item{label}{The label of the case/observation.}
  \item{cluster}{The id of the cluster to which the observation is assigned.}
  \item{path}{The sequence of child numbers (seperated by '.') at each level from top to the cluster.}
}
}
\description{
Output files of the Python script are read a to construct
a tibble with clusters which can be compared with a clustering from SCCA provided both clusterings are
based on the same dataset and axis.
}
\details{
This function is a helper for \code{\link{scca_py_overlap_test}}
}
