% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscca_bic.R
\name{gscca.bic}
\alias{gscca.bic}
\title{gscca model selection using bic}
\usage{
gscca.bic(x, y, edgex, edgey, maxsteps = 2000, plain = T,
  cv.method = "bic", gamma.u, gamma.v, lambda.u = NULL, lambda.v = NULL,
  Sx = NULL, Sy = NULL, ccor = F, thresh = 0.5)
}
\arguments{
\item{gamma, lambda:}{lambda*fused penalty+gamma*lambda*l1 penalty.}

\item{ccor:}{whether re-estiamte the coeeficient when doing BIC}
}
\value{
estimated coefficients, correlation and optimal parameters
}
\description{
gscca model selection using bic
}
