% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{scca}
\alias{scca}
\title{Spectral Clustering Correpondence Analysis.}
\usage{
scca(M)
}
\arguments{
\item{M}{A matrix representing a bi-partite graph. The rows and columns must be labeled}
}
\value{
\code{scca} returns a cluster tree as a recursive list.
}
\description{
The function \code{scca} performs a spectral clustering correspondence analysis on a given
bi-partite graph. It is a form a hierarchical clustering. On every level the sub-clustering is guided by
a spectral analysis to estimate the number of sub-clusters (the k of kmaens). See .....
}
\examples{
\dontrun{
data('carnivora', package = 'SCCAR')
scca(M = carnivora$M)
}
}
