% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCA.R
\docType{package}
\name{SCCA}
\alias{SCCA}
\title{Spectral Clustering Correspondence Analysis in R}
\description{
The SCCA package implements in R the methodological approach to CA as proposed
in \strong{A network view of correspondence analysis: applications to ecology and economic complexity}
(\href{https://www.uu.nl}{van Dam et al; 2021}).
}
\section{Exported functions}{

\itemize{
  \item \code{\link{scca_compute}}
  \item \code{\link{scca_get_clusters}}
  \item \code{\link{scca_get_node}}
  \item \code{\link{scca_plot_spectrum}}
  \item \code{\link{scca_print}}
  \item \code{\link{compute_rca}}
  \item \code{\link{eigengap_heuristic}}
}
}

\section{Exported data}{

\itemize{
  \item \code{\link{carnivora} \link{carnivora_sites} \link{carnivora_species}}
  \item \code{\link{exports} \link{exports_products} \link{exports_products}}
 }
}

