% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca.R
\name{scca}
\alias{scca}
\title{Spectral Clustering Correpondence Analysis.}
\usage{
scca(M)
}
\arguments{
\item{M}{A numeric matrix ......}
}
\value{
The function returns a list with .....
}
\description{
\code{scca} returns the ... of its argument M.
}
\details{
This function ....
}
\examples{
\dontrun{
data('carnivora', package = 'SCCA')
scca(M = carnivora$M)
}
}
