% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute.R
\name{scca_compute}
\alias{scca_compute}
\title{Spectral Clustering Correpondence Analysis.}
\usage{
scca_compute(m)
}
\arguments{
\item{m}{A matrix representing a bi-partite graph. The rows and columns must be labeled}
}
\value{
\code{scca} returns a cluster tree as a recursive list.
}
\description{
The function \code{scca} performs a spectral clustering correspondence analysis on a given
bi-partite graph. It is a form a hierarchical clustering. On every level the sub-clustering is guided by
a spectral analysis to estimate the number of sub-clusters (the k of kmaens). See .....
}
\examples{
\dontrun{
data('carnivora', package = 'sccar')
scca_compute(M = carnivora$M)
}
}
