% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_validity_test.R
\name{scca_compute_dist}
\alias{scca_compute_dist}
\title{Compute and Store Distance Matrix}
\usage{
scca_compute_dist(m, filename = NULL, method = "euclidian")
}
\arguments{
\item{m}{Incidence or adjacency matrix}

\item{filename}{String, name of the file in which to store the distance matrix. If NULL (default) the
distance will not be stored, but only returned}

\item{method}{See: \code{\link[stats]{dist}}}
}
\value{
The distance matrix as an \emph{dist} object
}
\description{
Compute and Store Distance Matrix
}
\details{
This function is a wrapper around \link[stats]{dist}.
Some validity and stability tests use a distance (or dissimilarity) matrixto calculate their measures (e.g. Dunn Index).
Calculating a distance matrix can take a long time (up to hours) depending on the number of observations (rows).
This function has the option to store the distance matrix in a .rds file.
}
