% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree.R
\name{scca_compute_tree}
\alias{scca_compute_tree}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree(
  labels,
  m,
  child,
  depth,
  n_node,
  iter.max = 10,
  nstart = 50,
  decomp = "svd",
  heuristic = eigengap_heuristic
)
}
\arguments{
\item{labels}{The labels (character vector) defining the (sub-)set of dataset m to be analyzed}

\item{m}{Dataset as an bipartite or incidence matrix}

\item{child}{The child number of this node within its siblings (= same parent)}

\item{depth}{The depth (integer) of this node in the tree}

\item{n_node}{Number of this node in the tree. This is a depth-first, pre-order numbering}

\item{iter.max}{The maximum number (integer) of iterations kmeans may take to compute the clusters}

\item{nstart}{The number of clusterings from which kmeans chooses the best}

\item{decomp}{The decomposition method to use: 'svd' (default) or 'svds'.}

\item{heuristic}{The function to use for calculating the number of expected clusters (k) and the embedding}
}
\description{
Build recursively a scca tree
}
