% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree.R
\name{scca_compute_tree}
\alias{scca_compute_tree}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree(
  labels,
  m,
  child,
  depth,
  max_depth,
  n_node,
  iter.max = 10,
  nstart = 50,
  max_eigenvalues = max_eigenvalues,
  decomp = "svd",
  heuristic = eigengap_heuristic
)
}
\arguments{
\item{labels}{The labels (character vector) defining the (sub-)set of data set m to be analyzed}

\item{m}{A matrix representing a bi-partite network. The matrix must have row names and column names.}

\item{child}{The child number of this node within its siblings (= nodes with same parent)}

\item{depth}{The depth (integer) of this node in the tree.}

\item{max_depth}{The maximum allowed depth of the analysis proces. If Inf (default) the analysis goes on untill a stop condition has been met.}

\item{n_node}{Number of the node in the tree. This is a depth-first, pre-order numbering, starting with 1 at the top node (a.k.a. root)}

\item{iter.max}{The maximum number of iterations \emph{kmeans} is allowed to make. Default is 10.}

\item{nstart}{Number of random cluster sets kmeans may choose to start with. Default is 25.}

\item{max_eigenvalues}{Restrict the number of computed eigenvalues to max_eigenvalues. The default is 25.}

\item{decomp}{The decomposition function to use. Choices are \emph{svd} (default) and \emph{svd}}

\item{heuristic}{The function to use for calculating the number of clusters. The default is \emph{eigengap_heuristic}}
}
\description{
See \code{\link{scca_compute}} for description
}
\details{
Function \emph{scca_compute_tree} calls itself recursively for every sub cluster found at a node until one of the  stop conditions
is met:
\itemize{
  \item{The heuristic finds only one prominent eigenvalue (k = 1)}
  \item{The maximum depth in a branch has been reached. \emph{k} will be set to 0}
  \item{The number of observations in the subset is too small to perform \emph{svd(s)}. This will raise a warning and \emph{k} will be set to -1}
}
}
