% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree.R
\name{scca_compute_tree}
\alias{scca_compute_tree}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree(labels, child, m, level, decomp_axis)
}
\arguments{
\item{labels}{Character vector; The labels defining the (sub-)cluster to be analyzed}

\item{child}{Numeric. Node number within its siblings (= same parent)}

\item{m}{Numeric matrix; Laplacian}

\item{level}{Integer; the depth of this node in the tree}

\item{decomp_axis}{Character string ('row', 'col)}
}
\description{
Build recursively a scca tree
}
\note{
Expects an matrix M accessible from within its environment
}
