% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree.R
\name{scca_compute_tree}
\alias{scca_compute_tree}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree(
  labels,
  m,
  child,
  depth,
  n_node,
  iter.max = 10,
  nstart = 50,
  decomp
)
}
\arguments{
\item{labels}{Character vector; The labels defining the (sub-)cluster to be analyzed}

\item{m}{Numeric matrix; Laplacian}

\item{child}{Numeric. Node number within its siblings (= same parent)}

\item{depth}{Integer the depth of this node in the tree}

\item{n_node}{Number of this node in the tree. This is a depth-first, pre-order numbering}

\item{iter.max}{Integer; the maximum number of iterations kmeans may take to compute the clusters}

\item{nstart}{Integer; the number of clusterings from which kmeans chooses the best}

\item{decomp}{The decomposition method to use: 'svd' or 'svds'.}
}
\description{
Build recursively a scca tree
}
