% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_compute_tree.R
\name{scca_compute_tree}
\alias{scca_compute_tree}
\title{Build recursively a scca tree}
\usage{
scca_compute_tree(m, labels, level, axis)
}
\arguments{
\item{m}{Numeric matrix; Laplacian}

\item{labels}{Character vector; The labels defining the rows/columns of the submatrix}

\item{level}{Integer; the depth of the node in the tree}

\item{axis}{Vector; should the rows or columns be subsetted}
}
\description{
Build recursively a scca tree
}
\note{
Expects an matrix M accessible from within its environment
}
