% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_get_clusters.R
\name{scca_get_clusters}
\alias{scca_get_clusters}
\title{Get Final Clusters from an SCCA Analysis}
\usage{
scca_get_clusters(scca)
}
\arguments{
\item{scca}{An SCCA analysis tree as output from a run of \code{\link{scca_compute}} on a dataset}
}
\value{
A tibble with 3 variables (columns)
\describe{
  \item{label}{The label of the case or observation}
  \item{cluster}{The id of the cluster to which the observation has been assigned. }
  \item{path}{The sequence of child numbers (seperated by '.') on the path from the top node to the leaf node}

}
}
\description{
Produces for every observation a record with the label of the observation
and the final (or leaf) cluster to which the observation has been assigned by scca_compute.
}
