% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_print.R
\name{scca_get_node}
\alias{scca_get_node}
\title{Retrieve Attributes of a Node in an SCCA tree}
\usage{
scca_get_node(scca, node)
}
\arguments{
\item{scca}{An SCCA tree}

\item{node}{The number of the node to retrieve the attributes from.}
}
\value{
List with 5 elements:
\describe{
  \item{labels}{The labels (rownames) defining the subset of observations at this node}
  \item{spectrum}{Sorted eigenvalues}
  \item{eigen_vec_1}{First eigenvector}
  \item{eigen_vec_2}{Second eigenvector}
  \item{eigen_vec_3}{Third eigenvector}
}
}
\description{
Retrieve Attributes of a Node in an SCCA tree
}
\details{
Returns NULL if \emph{node} doesn't exists.
}
