% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_print.R
\name{scca_plot_spectrum}
\alias{scca_plot_spectrum}
\title{Plot Spectrum of a Cluster}
\usage{
scca_plot_spectrum(scca, node_id, plot = TRUE)
}
\arguments{
\item{scca}{The output of \code{scca_compute}}

\item{node_id}{The node number (integer or coercible to) of the cluster in the output tree}

\item{plot}{If TRUE (default) the spectrum will be plotted. If FALSE only the spectrum data is returned.}
}
\description{
Plots the sorted spectrum of a (sub-)cluster at a specific node in the output tree of
\code{scca_compute}.
}
\details{
The node numbers can be displayed by \code{scca_print(scca)}. The function returns a tibble containing
  the spectrum data which can (e.g.) be used for your own plotting endeavours.
}
\examples{
  \dontrun{
     s <- scca_compute(t(carnivora))
     scca_print(scca = s, 'k', 'n_labs')
     scca_plot_spectrum(scca = s, node_id = '3')
}
}
