% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scca_stability_test.R
\name{scca_stability_test}
\alias{scca_stability_test}
\title{SCCA Stability Test}
\usage{
scca_stability_test(m, drop_vars = NULL)
}
\arguments{
\item{m}{Matrix representing an incidence (or bipartite) networ}

\item{drop_vars}{Integer vector, the index of the variables (columns) to be dropped (one by one) when testing
the stability. When omitted, all the columns are dropped one by one.}
}
\value{
A tibble with two columns:
\describe{
  \item{var_id}{Index of the dropped variable}
  \item{var_apo}{The average proportion of overlap between base clustering and clustering after dropping variable}
}
}
\description{
Tests the stability of an SCCA clustering. Only the clustering at the 'leaves' are considered. After a normal
clustering with \code{\link{scca_compute}} the clustering process is repeated while dropping one of variables from the user provided list.
This proces is repeated fro every variable in the list. Every time the clustering output is compared with the original clustering.
}
\details{
The stability measure is the average proportion of overlap (APO: the reverse of APN).
APO can be understood as the chance that if 2 observations are in the same cluster in the base clustering they are also
in the same cluster in the clustering with 1 column dropped.
}
