% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scca.R
\name{scca_write_analysis}
\alias{scca_write_analysis}
\title{Write a SCCA outcomne to file}
\usage{
scca_write_analysis(scca_tree, vec_name = "V_", spec_name = "S_")
}
\arguments{
\item{scca_tree}{Output tree of an call to scca_compute}

\item{vec_name}{Prefix of the files with the Eigenvectors}

\item{spec_name}{Prefix of files with the spectra}
}
\value{
Function returns TRUE
}
\description{
Write a SCCA outcomne to file
}
\details{
Writes the results of an SCCA analysis to .csv files. At every tree node the subcluster is is printed along
with its 3 most prominent Eigenvectors. The filename contains the path from the top to the tree node.
Another file contains the spectra found at this stage.
}
