% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scca.R
\name{scca_write_analysis}
\alias{scca_write_analysis}
\title{Write a SCCA outcomne to file}
\usage{
scca_write_analysis(scca_tree, vec_name = "V_", spec_name = "S_")
}
\arguments{
\item{scca_tree}{list, the tree resulting from a call to 'scca_compute'}

\item{vec_name}{character string; prefix of the files with the clusters and the Eigenvectors}

\item{spec_name}{chaacter string; prefix of files with the spectra}
}
\value{
TRUE
}
\description{
Write a SCCA outcomne to file
}
\details{
Writes the results of an SCCA analysis to .csv files. At every tree node two files are written.
1. The first column contains labels of the subcluster and columns 2 to 4 are the 3 most prominent Eigenvectors

2. The first column contains the spectrum of the subcluster and the second column the explained variance

The filename is a concatenation of the prefix and de path from the top to the node.
}
