% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccar.R
\docType{package}
\name{sccar}
\alias{sccar}
\title{An package for Spectral Clustering Correspondence Analysis in R}
\description{
SCCA is a methode for ....................
}
\section{Exported functions}{

\itemize{
  \item \code{\link{scca_compute}}
  \item \code{\link{scca_compute_dist}}
  \item \code{\link{scca_get_clusters}}
  \item \code{\link{scca_get_node}}
  \item \code{\link{scca_overlap_test}}
  \item \code{\link{scca_plot_spectrum}}
  \item \code{\link{scca_print}}
  \item \code{\link{scca_py_overlap_test}}
  \item \code{\link{scca_stability_test}}
  \item \code{\link{scca_validity_test}}
  \item \code{\link{scca_write_clustering}}
 }
}

