% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAR.R
\docType{package}
\name{SCCAR}
\alias{SCCAR}
\title{SCCAR: A package for Spectral Clustering Corresponce Analysis.}
\description{
The SCCARR package provides functions for:
}
\section{SCCAR functions}{

The  functions ...
}

