% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccar.R
\docType{package}
\name{sccar}
\alias{sccar}
\title{sccar: A package for Spectral Clustering Corresponce Analysis.}
\description{
The sccar package provides R functions for ....
}
\section{sccar functions}{

The exported functions ...
}

